refactor: mcp_client
